Add 'cd' command to enter a castle
The command will enter home/ if it exists, or else the repo root.

This requires the user to source a helper script (homeshick.{sh,csh})
that performs the 'cd' command within the context of the current shell.

If the main homeshick script is called with a 'cd' command, this just
displays the normal 'cd' help message, which indicates that homeshick
must be invoked via the helper script for the 'cd' command to work.