Fix parse_url for repos with dots in name
The regex that was being used in parse_url searched until a slash (/) or
dot (.) was found and then looked for an optional (.git) at the end.
This caused problems with repo names like vim-scripts/ebnf.vim.git and
actually caused the clone command to put the repo into a strange
"git:/<user>/<repo>" directory structure.

Fix replaces complicate sed regex and associated checks with 'basename'