Add --git option to clone command
When using the shorthand notation for the clone command (homeshick clone
<user>/<repo>), homeshick converts this to a GitHub repo with the https
protocol (https://github.com/<user>/<repo>.git).

This can be inconvenient if it is a private repo, as you then have to
type your password. For such cases, using the git protocol might be
preferred.

This commit introduces a --git option that, when used, uses the Git
protocol instead: git@github.com:<user>/<repo>.git

That option is ignored for other commands or full repo addresses.